Get mongo to tell us the replica set name when connecting locally
socket.gethostname() won't always match the local name of the replica
set member; one example is when gethostname() returns a shortname, but
the replica set member name is an FQDN.

In a previous version of the code (before 32409ca0), we asked mongo to
tell us which member we were, by querying the "me" key from the hash
returned by the "ismaster" command.  This seems like the Right Thing to
do -- rather than guess what name we have, just get mongo to tell us.